+++ EstimoteIndoorLocationSDK/Headers/EILColor.h
@@ -1,0 +1,22 @@
+//  Copyright © 2017 Estimote. All rights reserved.
+
+#import <Foundation/Foundation.h>
+
+/**
+ * Possible colors of Estimote devices.
+ */
+typedef NS_ENUM(NSInteger, EILColor)
+{
+    EILColorUnknown,
+    EILColorMintCocktail,
+    EILColorIcyMarshmallow,
+    EILColorBlueberryPie,
+    EILColorSweetBeetroot,
+    EILColorCandyFloss,
+    EILColorLemonTart,
+    EILColorWhite,
+    EILColorBlack,
+    EILColorCoconutPuff,
+    EILColorTransparent
+};
+