+++ EstimoteIndoorLocationSDK/Headers/EILIndoorSDK.h
@@ -1,7 +1,7 @@
 // Copyright (c) 2017 Estimote. All rights reserved.
 
 /** Version of this iOS Indoor Location SDK. */
-#define EIL_IOS_INDOOR_SDK_VERSION_STRING @"2.7.1"
+#define EIL_IOS_INDOOR_SDK_VERSION_STRING @"3.0.0-alpha.1"
 
 
 /** Library header, import this to import all of the public types in the Indoor SDK. */
@@ -13,12 +13,14 @@
 #import "EILBackgroundIndoorLocationManager.h"
 
 // Location data structures.
+#import "EILColor.h"
 #import "EILPoint.h"
 #import "EILOrientedPoint.h"
 #import "EILOrientedLineSegment.h"
 #import "EILLocationLinearObject.h"
 #import "EILPositionedBeacon.h"
 #import "EILLocation.h"
+#import "EILLocationPin.h"
 
 // Manually building location.
 #import "EILLocationBuilder.h"
@@ -35,3 +37,9 @@
 #import "EILRequestFetchPublicLocations.h"
 #import "EILRequestRemoveLocation.h"
 #import "EILRequestModifyLocation.h"
+
+// Location Pin management
+#import "EILRequestAddPinToLocation.h"
+#import "EILRequestFetchLocationPins.h"
+#import "EILRequestUpdateLocationPin.h"
+#import "EILRequestRemoveLocationPin.h"