+++ EstimoteIndoorLocationSDK/Headers/EILRequestFetchPublicLocations.h
@@ -1,13 +1,12 @@
-//  Copyright © 2015 Estimote. All rights reserved.
+//  Copyright © 2017 Estimote. All rights reserved.
 
 #import <Foundation/Foundation.h>
-#import <EstimoteSDK/EstimoteSDK.h>
 
 @class EILLocation;
 
 NS_ASSUME_NONNULL_BEGIN
 
-/** 
+/**
  * A block object to be executed when the request finishes.
  */
 typedef void(^EILRequestFetchPublicLocationsBlock)(NSArray<EILLocation *> * _Nullable locations, NSError * _Nullable error);
@@ -18,7 +17,7 @@ typedef void(^EILRequestFetchPublicLocationsBlock)(NSArray<EILLocation *> * _Nul
  * All the locations that contain at least one of the given beacons (identified
  * by their identifiers) will be fetched.
  */
-@interface EILRequestFetchPublicLocations : ESTRequestPostJSON
+@interface EILRequestFetchPublicLocations : NSObject
 
 /**
  * Returns a new request object for fetching public locations for given beacons.
@@ -27,16 +26,6 @@ typedef void(^EILRequestFetchPublicLocationsBlock)(NSArray<EILLocation *> * _Nul
  * @return A request initialized with identifiers of beacons.
  */
 - (instancetype)initWithBeaconIdentifiers:(NSArray<NSString *> *)identifiers;
-
-/**
- * Returns a new request object for fetching public locations for given beacons.
- *
- * @param beaconMacs An array of `NSString *` containing mac addresses of beacons that are used for fetching nearby public locations.
- * @return A request initialized with mac addresses of beacons.
- *
- * This method is deprecated. Use initWithBeaconIdentifiers: instead.
- */
-- (instancetype)initWithBeaconMacs:(NSArray<NSString *> *)beaconMacs __deprecated_msg("Use initWithBeaconIdentifiers: instead.");
 
 /**
  * Sends request to Estimote Cloud with completion block.