+++ EstimoteIndoorLocationSDK/Headers/EILRequestRemoveLocation.h
@@ -1,13 +1,12 @@
-//  Copyright © 2015 Estimote. All rights reserved.
+//  Copyright © 2017 Estimote. All rights reserved.
 
 #import <Foundation/Foundation.h>
-#import <EstimoteSDK/EstimoteSDK.h>
 
 @class EILLocation;
 
 NS_ASSUME_NONNULL_BEGIN
 
-/** 
+/**
  * A block object to be executed when the request finishes.
  */
 typedef void(^EILRequestRemoveLocationBlock)(EILLocation * _Nullable location, NSError * _Nullable error);
@@ -20,7 +19,7 @@ typedef void(^EILRequestRemoveLocationBlock)(EILLocation * _Nullable location, N
  * You can find your API App ID and API App Token in the Apps: http://cloud.estimote.com/#/apps
  * section of the Estimote Cloud: http://cloud.estimote.com/.
  */
-@interface EILRequestRemoveLocation : ESTRequestPostJSON
+@interface EILRequestRemoveLocation : NSObject
 
 /**
  * Returns a new request object for removing location from Estimote Cloud.
@@ -31,12 +30,20 @@ typedef void(^EILRequestRemoveLocationBlock)(EILLocation * _Nullable location, N
 - (instancetype)initWithLocation:(EILLocation *)location;
 
 /**
+ * Returns a new request object for removing location from Estimote Cloud.
+ *
+ * @param locationIdentifier Identifier of the location to be removed from Estimote Cloud.
+ * @return A request initialized with location.
+ */
+- (instancetype)initWithLocationIdentifier:(NSString *)locationIdentifier;
+
+/**
  * Sends request to Estimote Cloud with completion block.
  *
  * param completion Completion block to be executed when the request finishes.
  */
 - (void)sendRequestWithCompletion:(EILRequestRemoveLocationBlock)completion;
 
-@end
-
 NS_ASSUME_NONNULL_END
+
+@end